// src/enums/ETokenType.ts
export enum ETokenType {
  OPEN_OBJECT = "OPEN_OBJECT",
  CLOSE_OBJECT = "CLOSE_OBJECT",
  DEFINE_VALUE = "DEFINE_VALUE",
  NEXT_LINE = "NEXT_LINE",
  OPEN_KEY = "OPEN_KEY",
  CLOSE_KEY = "CLOSE_KEY",
  KEY = "KEY",
  OPEN_RULES = "OPEN_RULES",
  CLOSE_RULES = "CLOSE_RULES",
  RULES = "RULES",
  OPEN_ARRAY = "OPEN_ARRAY",
  CLOSE_ARRAY = "CLOSE_ARRAY",
}